session/requirements.txt
asgiref==3.6.0
boto3==1.28.11
botocore==1.31.11
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.2
defusedxml==0.7.1
dj-rest-auth==4.0.1
Django==4.1.7
django-allauth==0.54.0
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
jmespath==1.0.1
mysql-connector-python==8.1.0
mysqlclient==2.2.0
oauthlib==3.2.2
Pillow==10.0.0
protobuf==4.21.12
pycparser==2.21
PyJWT==2.7.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.16
